renames getAllDataGroupActions to getAllGroupMgmtDataAction.js
